pyproject.toml
[project]
name = "ml-serving-rest-rpc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "grpcio-tools>=1.71.0",
    "mypy-protobuf>=3.6.0",
    "numpy>=2.2.5",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
]
